// old-pixi/src/types/grid.ts
export type Point = [number, number];

/**
 * Describes the type of cell, either a land cell for a landmass or a water cells from the water next to a landmass.
 */
export enum CellType {
  Highland = 2,
  Land = 1,
  Water = -1,
}

/**
 * Describes the type of a feature.
 */
export enum FeatureType {
  INVALID = 'invalid',
  ISLAND = 'island',
  OCEAN = 'ocean',
  LAKE = 'lake',
}
// ...
export interface Feature {
  index: number;
  isLand?: boolean;
  isBorder?: boolean;
  type: FeatureType;
}
// ...
export interface River {
  index: number;

  /**
   * The cell source of this river, where this river originated.
   */
  source: number;

  /**
   * The mouth of the river, which is one cell off of the final river cell, which should usually be
   * a water cell.
   */
  mouth: number;

  /**
   * How much water this river discharges into its output water feature in m³ per second.
   */
  discharge: number;

  /**
   * The length of the river from the source to its mouth.
   */
  length: number;

  /**
   * The average width of this rivers throughout its cells.
   */
  width: number;

  /**
   * How much this rivers grows or shrinks as you navigate up or down the river.
   */
  widthFactor: number;

  /**
   * The width of the source of this river, usually 0 if it originates from a land cell.
   */
  sourceWidth: number;

  /**
   * The id of the river's parent if any. The ID will be 0 if the river has no parent (it's a root river).
   */
  parent: number;

  /**
   * The list of cells that make up this river. The river is configured to meander throughout its cells to give the
   * appearance of a real river and not a straight line.
   */
  cells: number[];
}

/**
 * voronoi cells generated from the voronoi algorithm.
 */
export interface Cells {
  /**
   * Array of indexes for each cell in the grid.
   */
  indexes: Uint32Array | Uint8Array | Uint16Array;

  /**
   * Cell vertices for the cell at the given index.
   */
  vertices: number[][];

  /**
   * Cells adjacent to the cell at the given index.
   */
  adjacentCells: number[][];

  /**
   * List of all the cells, identifying which one is a near border cell or not.
   */
  nearBorderCells: boolean[];

  /**
   * Array of heights from the heightmap for each cell.
   */
  heights: Uint32Array | Uint8Array | Uint16Array;

  /**
   * Array of features indexes for individual cells, can be used to query the actual feature from the features array of
   * the grid.
   */
  features: Uint16Array;

  /**
   * Array of types for cells to identify which cells are part of the land/coast of a landmass and which cells are the
   * water near the landmass.
   */
  types: CellType[];

  /**
   * Array of temperatures for the given cell IDs, generated by creating a random position for the map on a globe.
   */
  temperatures: Int8Array;

  /**
   * Array of precipitation for the given cell IDs, used to generate biomes for individual cells as it tells how dry or
   * wet an area of the map is.
   */
  precipitation: Uint8Array;

  /**
   * Array of water flux values for each cell, useful to generate rivers and biomes, in m³ per second.
   */
  waterFlux: Uint16Array;

  /**
   * Array of river indexes for a cell. 0 if there are no rivers on that cell, otherwise it defines the ID for the river
   * from the grid object.
   */
  rivers: Uint16Array;

  /**
   * Array of confluence for rivers, this describes when two or more rivers merge into one another.
   */
  confluences: Uint8Array | Uint16Array;
}
// ...
export interface Vertices {
  /**
   * Coordinates for the vertices of the cells at the given index, will be set to the center of the triangle generated
   * by delaunator.
   */
  coordinates: Point[];

  /**
   * Neighbouring vertices for the vertex at the given index.
   */
  neighbours: number[][];

  /**
   * Adjacent cells of the vertex at the given index.
   */
  adjacent: [number, number, number][];
}

/**
 * The definition for a map's Grid, which contains all the cells and vertices definitions, but the description
 * of what kind of map feature they contain (heightmap, ocean, lake, river, etc.).
 */
export interface Grid {
  spacing: number;
  boundary: Point[];
  points: Point[];
  cellsX: number;
  cellsY: number;

  cells: Cells;
  vertices: Vertices;
  features: Feature[];
  rivers: River[];
}